msv: fix problem validating a choice with multiple list children
ResidualCalculator.calcResidual(): recursive calls to this function
overwrite the "this.token" member, so that after the inner
calcResidual() returns, the outer call will continue with a sub-string
of its own token.

For example, StringToken.match(ListExp) is called with resCalc.token =
"top left" and on return from the first call to resCalc.calcResidual,
resCalc.token = "top" but there was no match, so immediately return
false; the next call to StringToken.match(ListExp) will have
resCalc.token = "top" and fail to match sequence of "top" "left".